# api/openapi-lock.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "org.quicko.qpon",
    "description": "Qpon is a coupon framework built by Quicko. Built in a scalable, multi-tenant thought process, the Qpon framework enables businesses to do campaigns and run offers using the Qpon API.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://{{quicko_host}}"
    }
  ],
...
  "paths": {
...
    "/offers": {
      "get": {
        "tags": [
          "Offers"
        ],
        "summary": "Fetch offers",
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "external_item_id",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "example": "5126317000000116190"
          },
          {
            "name": "external_customer_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to get customer specific coupon codes",
            "example": "vishvajit@quicko.com"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to sort ascending or descending",
            "example": "desc"
          },
          {
            "name": "discount_type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "percent"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page offset. Default is 0",
            "example": "0"
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page size. Default is 10",
            "example": "10"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/offer": {
      "get": {
        "tags": [
          "Offers"
        ],
        "summary": "Fetch offer",
        "parameters": [
          {
            "name": "x-auth-username",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Username for identification",
            "example": "{{x-auth-username}}"
          },
          {
            "name": "x-auth-password",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Password for authentication",
            "example": "{{x-auth-password}}"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Organization id",
            "example": "{{organization_id}}"
          },
          {
            "name": "external_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Optional filter to get customer specific offer",
            "example": "quickoall01"
          },
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "BAJAJFIN10"
          },
          {
            "name": "item_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "all"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/{organization_id}/items": {
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Create item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "item"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "custom_fields": {
                    "type": "object",
                    "properties": {
                      "key": {
                        "type": "string"
                      },
                      "value": {
                        "type": "string"
                      }
                    }
                  }
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Fetch items",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page offset. Default is 0",
            "example": "0"
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page size. Default is 10",
            "example": "10"
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "example": "test"
          },
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/{organization_id}/items/{item_id}": {
      "patch": {
        "tags": [
          "Items"
        ],
        "summary": "Update item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "description": "test description"
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          },
          {
            "name": "item_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{item_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Items"
        ],
        "summary": "Delete item",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          },
          {
            "name": "item_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{item_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Fetch item",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          },
          {
            "name": "item_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{item_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/{organization_id}/customers": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Create customer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "@entity": {
                    "default": "customer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "external_id": {
                    "type": "string"
                  }
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Fetch customers",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page offset. Default is 0",
            "example": "0"
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Optional page size. Default is 10",
            "example": "10"
          },
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/{organization_id}/customers/{customer_id}": {
      "patch": {
        "tags": [
          "Customers"
        ],
        "summary": "Update customer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "email": "dan@brown.com"
                }
              },
              "examples": {}
            }
          }
        },
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          },
          {
            "name": "customer_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{customer_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Customers"
        ],
        "summary": "Delete customer",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{organization_id}}"
          },
          {
            "name": "customer_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "{{customer_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
...
  }
}